move shared functions to utils
