+++ jiant/tasks/evaluate/core.py
@@ -1,7 +1,5 @@
 import collections
 import json
-import re
-import string
 
 from dataclasses import dataclass
 
@@ -19,6 +17,7 @@ import jiant.tasks.lib.templates.squad_style.utils as squad_style_utils
 import jiant.tasks.lib.mlqa as mlqa_lib
 from jiant.utils.python.datastructures import ExtendedDataClassMixin
 from jiant.utils.python.io import read_json
+from jiant.utils.string_comparing import f1_score, exact_match_score
 
 
 @dataclass
@@ -122,49 +121,6 @@ class SpanPredictionF1andEMScheme(BaseEvaluationScheme):
         return accumulator.get_accumulated()
 
     def compute_metrics_from_preds_and_labels(self, preds, labels):
-        def normalize_answer(s):
-            """Lower text and remove punctuation, articles and extra whitespace.
-            From official ReCoRD eval script
-            """
-
-            def remove_articles(text):
-                return re.sub(r"\b(a|an|the)\b", " ", text)
-
-            def white_space_fix(text):
-                return " ".join(text.split())
-
-            def remove_punc(text):
-                exclude = set(string.punctuation)
-                return "".join(ch for ch in text if ch not in exclude)
-
-            def lower(text):
-                return text.lower()
-
-            return white_space_fix(remove_articles(remove_punc(lower(s))))
-
-        def f1_score(prediction, ground_truth):
-            """Compute normalized token level F1
-            From official ReCoRD eval script
-            """
-            prediction_tokens = normalize_answer(prediction).split()
-            ground_truth_tokens = normalize_answer(ground_truth).split()
-            common = collections.Counter(prediction_tokens) & collections.Counter(
-                ground_truth_tokens
-            )
-            num_same = sum(common.values())
-            if num_same == 0:
-                return 0
-            precision = 1.0 * num_same / len(prediction_tokens)
-            recall = 1.0 * num_same / len(ground_truth_tokens)
-            f1 = (2 * precision * recall) / (precision + recall)
-            return f1
-
-        def exact_match_score(prediction, ground_truth):
-            """Compute normalized exact match
-            From official ReCoRD eval script
-            """
-            return normalize_answer(prediction) == normalize_answer(ground_truth)
-
         em = sum([exact_match_score(s1, s2) for s1, s2 in zip(preds, labels)]) / len(labels)
         f1 = sum([f1_score(s1, s2) for s1, s2 in zip(preds, labels)]) / len(labels)
         scores = {"f1": f1, "em": em, "avg": (f1 + em) / 2}
@@ -447,11 +403,11 @@ class ReCordEvaluationScheme(BaseEvaluationScheme):
             pred_ans = relevant_examples[psg_qns_pred].entity_str
 
             # F1
-            f1 = cls.metric_max_over_ground_truths(cls.f1_score, pred_ans, golds)
+            f1 = cls.metric_max_over_ground_truths(f1_score, pred_ans, golds)
             f1_ls.append(f1)
 
             # EM
-            em = cls.metric_max_over_ground_truths(cls.exact_match_score, pred_ans, golds)
+            em = cls.metric_max_over_ground_truths(exact_match_score, pred_ans, golds)
             em_ls.append(em)
             predictions_dict[psq_qns_idx] = psg_qns_pred
 
@@ -464,50 +420,6 @@ class ReCordEvaluationScheme(BaseEvaluationScheme):
         }
         metrics = Metrics(major=minor["f1_em"], minor=minor,)
         return predictions_dict, metrics
-
-    @classmethod
-    def normalize_answer(cls, s):
-        """Lower text and remove punctuation, articles and extra whitespace.
-        From official ReCoRD eval script
-        """
-
-        def remove_articles(text):
-            return re.sub(r"\b(a|an|the)\b", " ", text)
-
-        def white_space_fix(text):
-            return " ".join(text.split())
-
-        def remove_punc(text):
-            exclude = set(string.punctuation)
-            return "".join(ch for ch in text if ch not in exclude)
-
-        def lower(text):
-            return text.lower()
-
-        return white_space_fix(remove_articles(remove_punc(lower(s))))
-
-    @classmethod
-    def f1_score(cls, prediction, ground_truth):
-        """Compute normalized token level F1
-        From official ReCoRD eval script
-        """
-        prediction_tokens = cls.normalize_answer(prediction).split()
-        ground_truth_tokens = cls.normalize_answer(ground_truth).split()
-        common = collections.Counter(prediction_tokens) & collections.Counter(ground_truth_tokens)
-        num_same = sum(common.values())
-        if num_same == 0:
-            return 0
-        precision = 1.0 * num_same / len(prediction_tokens)
-        recall = 1.0 * num_same / len(ground_truth_tokens)
-        f1 = (2 * precision * recall) / (precision + recall)
-        return f1
-
-    @classmethod
-    def exact_match_score(cls, prediction, ground_truth):
-        """Compute normalized exact match
-        From official ReCoRD eval script
-        """
-        return cls.normalize_answer(prediction) == cls.normalize_answer(ground_truth)
 
     @classmethod
     def metric_max_over_ground_truths(cls, metric_fn, prediction, ground_truths):
